Fixed menu ordering
Turns out a value of the form NNNN.NN gets parsed as a number and not a string.
Adding a day to get the form NNNN.NN.NN resulted in an error, so I had to switch
to something that actually parses as a string.

Also renamed 'circa' to 'order' to make it clear what the field is used for.